archive/requirements.txt
# Power Sensor Simulator Dependencies
paho-mqtt==1.6.1
